Implemented method for adding new article
* Modified Artikal.java and ArtikalDTO.java classes.
* Frontend validation of form for adding new article.
* Created method for adding new article on a backend.
* Linked server and client with an AJAX call.
* Modified restaurantInfo.js: showing restaurant's articles instead of the dummy ones.

+++ src/main/java/dto/ArtikalDTO.java
@@ -4,7 +4,7 @@ import beans.TipArtikla;
 
 public class ArtikalDTO {
 	private String naziv;
-	private int cena;
+	private float cena;
 	private String tip;
 	private String restoran;
 	private int kolicina;
@@ -22,11 +22,11 @@ public class ArtikalDTO {
 		this.naziv = naziv;
 	}
 
-	public int getCena() {
+	public float getCena() {
 		return cena;
 	}
 
-	public void setCena(int cena) {
+	public void setCena(float cena) {
 		this.cena = cena;
 	}
